+++ Game/Code/Game/Entity/Components/EntityState.cs
@@ -4,7 +4,7 @@ namespace Daikon.Game;
 
 public record EntityState
 {
-    public int TimeStamp { get; init; }
+    public double TimeStamp { get; init; }
     public Vector3 Position { get; init; }
     public Vector3 Rotation { get; init; }
 }
